Move data-store helpers to 'helpers' folder
